🔧 Fix SCMS build by moving cron scopes to scms-core
CronEndpointScopes in the system.cron client bundle pulled the full
scms-server graph (including Node crypto). Also fix route test typing
and lint issues in cron/drain routes.
