+++ packages/scms-server/src/backend/cron/scopes.ts
@@ -1,5 +1,12 @@
-import { error401 } from '@curvenote/scms-core';
+import {
+  CronEndpointScopes,
+  type CronEndpointScope,
+  cronEndpointScope,
+  error401,
+} from '@curvenote/scms-core';
 import { type SignHandshakeTokenClaims, verifyHandshakeToken } from '../sign.handshake.server.js';
+
+export { CronEndpointScopes, type CronEndpointScope, cronEndpointScope };
 
 export type CronHandshakeConfig = {
   api: {
@@ -7,20 +14,6 @@ export type CronHandshakeConfig = {
     handshakeSigningSecret: string;
   };
 };
-
-/** Built-in cron callback endpoint scopes ({METHOD}:{path}). */
-export const CronEndpointScopes = {
-  JOB_QUEUE_DRAIN: 'POST:/v1/jobs/push-to-drain',
-  PROMOTE_SCHEDULED: 'POST:/v1/jobs/promote-scheduled',
-  TEXT_INTEGRITY_RETRY_SWEEP: 'POST:/v1/hooks/text-integrity/retry-sweep',
-} as const;
-
-export type CronEndpointScope = (typeof CronEndpointScopes)[keyof typeof CronEndpointScopes];
-
-export function cronEndpointScope(method: string, path: string): string {
-  const normalizedPath = path.startsWith('/') ? path : `/${path}`;
-  return `${method.toUpperCase()}:${normalizedPath.split('?')[0]}`;
-}
 
 export function assertEndpointScope(
   claims: SignHandshakeTokenClaims | undefined,
